WIP2 Add query logic for likelihood and significance
