+++ app/voting/serializers/__init__.py
@@ -3,3 +3,4 @@ from .ballot_option import BallotOptionSerializer
 from .room import RoomSerializer
 from .voting_session import VotingSessionSerializer
 from .user_vote import UserVoteSerializer
+from .suggestion import SuggestionSerializer
