+++ app/voting/serializers/suggestion.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+
+
+class SuggestionSerializer(serializers.Serializer):
+    id = serializers.IntegerField()
+    likelihood = serializers.FloatField()
+    significance = serializers.FloatField()
