+++ app/voting/views/__init__.py
@@ -3,3 +3,4 @@ from .ballot_option import BallotOptionViewSet
 from .room import RoomViewSet
 from .voting_session import VotingSessionViewSet
 from .user_vote import UserVoteViewSet
+from .suggestion import SuggestionViewSet
