+++ app/voting/views/suggestion.py
@@ -1,4 +1,5 @@
-from django.db.models import Q, F, Avg
+from django.db.models import Q, F, Avg, OuterRef, Subquery
+from django.db.models.functions import Abs
 from rest_framework import viewsets
 
 from voting.models import BallotOption, OptionCorrelation
@@ -15,7 +16,7 @@ class SuggestionViewSet(viewsets.ReadOnlyModelViewSet):
 
         return BallotOption.objects.exclude(
             # Do not offer suggestions that have already been voted on
-            user_vote__session=session_token
+            uservote__session=session_token
         ).annotate(
             likelihood=self.get_likelihood_annotation(session_token),
             significance=self.get_likelihood_annotation(session_token),
@@ -23,15 +24,15 @@ class SuggestionViewSet(viewsets.ReadOnlyModelViewSet):
             'id',
             'likelihood',
             'significance'
-        )
+        ).order_by('-likelihood')
 
-    def get_likelihood_annotation(session_token):
+    def get_likelihood_annotation(self, session_token):
         # Average the correlation value over all rows for predicates this session has not excluded
         # (i.e., either has not voted on or has voted matching the row)
         return Avg(
             'correlation_target__correlation',
-            filter=Q(
-                correlation_target__predicate__uservote__session__ne=session_token
+            filter=~Q(
+                correlation_target__predicate__uservote__session=session_token
             ) | Q(
                 correlation_target__predicate__uservote__session=session_token,
                 correlation_target__predicate_polarity=F(
@@ -40,17 +41,36 @@ class SuggestionViewSet(viewsets.ReadOnlyModelViewSet):
             )
         )
 
-    def get_significance_annotation(session_token):
+    def get_significance_annotation(self, session_token):
         # Calculate the absolute difference in correlation of the two polarities for a given target.
         # Exclude any correlation that this session has voted for, either as predicate or target.
         # (Exclude already voted predicates as they cannot be suggestions, and already
         # voted targets as we want to maximize effect on future suggestions)
         # e.g., if A has a 0.3 correlation with B, and ~A has a 0.8 correlation with B, then A has
         # a significance of 0.5 with B
+        correlation_change = OptionCorrelation.objects.filter(
+            predicate=OuterRef('pk'),
+            polarity=True
+        ).annotate(
+            correlation_false=Subquery(
+                OptionCorrelation.objects.filter(
+                    predicate=OuterRef('predicate'),
+                    polarity=False,
+                    target=OuterRef('target')
+                ).values('correlation')
+            )
+        ).values(
+            correlation_change=Abs(
+                F('correlation')-F('correlation_false')
+            )
+        )
         return Avg(
-            'correlation_predicate__correlation',
+            correlation_change,
             filter=Q(
-                correlation_predicate__predicate__uservote__session__ne=session_token,
-                correlation_predicate__target__uservote__session__ne=session_token
+                ~Q(
+                    correlation_predicate__predicate__uservote__session=session_token) &
+                ~Q(
+                    correlation_predicate__target__uservote__session=session_token
+                )
             )
         )
