+++ app/voting/views/utils.py
